refactor(ai): extract chat handler formatters into handler/formatters/ (U1)
Move ~2150 LOC of deterministic tool-response formatters out of the
6439-LOC handler.ts into a focused formatters/ subdirectory. Pure
mechanical move — no behavior change.

- handler/formatters/connections.ts — suggest_connections + helpers
- handler/formatters/reads.ts — list_*/get_*/suggest_mentors/navigation
- handler/formatters/prepares.ts — prepare_*/revoke + revised event
- handler/formatters/schedules.ts — extract_schedule_pdf
- handler/formatters/index.ts — dispatcher + error formatter + shared

handler.ts re-exports the symbols tests + executor still import; shim
removed in U7 once route.ts and tests target ./handler directly.

handler.ts: 6439 → 4289 LOC.
