+++ src/app/api/ai/[orgId]/chat/handler.ts
@@ -103,6 +103,28 @@ import {
   finalizeStageTimings,
 } from "@/lib/ai/chat-telemetry";
 import { aiLog, type AiLogContext } from "@/lib/ai/logger";
+import {
+  getNonEmptyString,
+  hasPendingConnectionDisambiguation,
+  looksLikeConnectionDisambiguationReply,
+  collectPhoneNumberFields,
+  formatDeterministicToolResponse,
+  formatDeterministicToolErrorResponse,
+  formatRevisedPendingEventResponse,
+  resolveHideDonorNamesPreference,
+  CONNECTION_PASS2_TEMPLATE,
+} from "./handler/formatters/index";
+
+export {
+  CONNECTION_PASS2_TEMPLATE,
+  collectPhoneNumberFields,
+  formatSuggestConnectionsResponse,
+  formatDeterministicToolResponse,
+  formatDeterministicToolErrorResponse,
+  formatRevisedPendingEventResponse,
+  resolveHideDonorNamesPreference,
+  type DonationResponseOptions,
+} from "./handler/formatters/index";
 
 export interface ChatRouteDeps {
   createClient?: typeof createClient;
@@ -254,22 +276,6 @@ function looksLikeStructuredJobDraft(message: string): boolean {
   return hasJobContext && structuredFieldMatches >= 2 && message.trim().length >= 80;
 }
 
-export const CONNECTION_PASS2_TEMPLATE = [
-  "CONNECTION ANSWER CONTRACT:",
-  "- If suggest_connections returned state=resolved, respond using this exact shape:",
-  "  Top connections for [source person name]",
-  "  1. [suggestion name] - [subtitle if present]",
-  "  Why: [reason], [reason], [reason]",
-  "- Use at most 3 suggestions.",
-  "- Use only the returned source_person, suggestions, subtitles, and normalized reason labels.",
-  "- Do not mention scores, UUIDs, Falkor, SQL fallback, freshness, or internal tool details.",
-  "- Do not add a concluding summary sentence.",
-  "- If state=ambiguous, ask the user which returned option they mean.",
-  "- If the user's next message picks one of the ambiguous options you previously listed (by name, subtitle, or position), call suggest_connections again using the person_type and person_id from the [ref: person_type:person_id] tag on the matching line, not person_query.",
-  "- If state=not_found, say you couldn't find that person in the organization's member or alumni data and ask for a narrower identifier.",
-  "- If state=no_suggestions, say you found the person but there is not enough strong professional overlap yet to recommend a connection.",
-].join("\n");
-
 const CONNECTION_PASS1_DISAMBIGUATION_INSTRUCTION = [
   "CONNECTION TOOL ROUTING:",
   "- If the latest assistant message listed ambiguous suggest_connections options with [ref: person_type:person_id] tags and the user replies with a choice by number, position, name, or subtitle, call suggest_connections again.",
@@ -308,10 +314,6 @@ class ToolGroundingVerificationError extends Error {
   }
 }
 
-interface SuggestConnectionDisplayReason {
-  label?: unknown;
-}
-
 function buildSseResponse(stream: ReadableStream<Uint8Array>, headers: HeadersInit, threadId: string) {
   return new Response(stream, {
     headers: {
@@ -319,43 +321,6 @@ function buildSseResponse(stream: ReadableStream<Uint8Array>, headers: HeadersIn
       "x-ai-thread-id": threadId,
     },
   });
-}
-
-interface SuggestConnectionDisplayRow {
-  name?: unknown;
-  subtitle?: unknown;
-  reasons?: unknown;
-}
-
-interface SuggestConnectionDisplayPayload {
-  state?: unknown;
-  source_person?: { name?: unknown } | null;
-  suggestions?: unknown;
-  disambiguation_options?: unknown;
-}
-
-interface AnnouncementDisplayRow {
-  title?: unknown;
-  published_at?: unknown;
-  audience?: unknown;
-  is_pinned?: unknown;
-  body_preview?: unknown;
-}
-
-interface NavigationDisplayTarget {
-  label?: unknown;
-  href?: unknown;
-  description?: unknown;
-  kind?: unknown;
-  userCanAccess?: unknown;
-  manualSteps?: unknown;
-  assistantCanHelpWith?: unknown;
-}
-
-interface NavigationDisplayPayload {
-  state?: unknown;
-  query?: unknown;
-  matches?: unknown;
 }
 
 interface PendingActionToolPayload {
@@ -433,1493 +398,6 @@ async function listPendingEventActionsForThread(
       row.payload != null &&
       typeof row.payload === "object"
   );
-}
-
-function getNonEmptyString(value: unknown): string | null {
-  return typeof value === "string" && value.trim().length > 0 ? value.trim() : null;
-}
-
-function formatIsoDate(value: unknown): string | null {
-  const iso = getNonEmptyString(value);
-  return iso ? iso.slice(0, 10) : null;
-}
-
-function formatDisplayRow(row: { name?: unknown; subtitle?: unknown }): string | null {
-  const name = getNonEmptyString(row.name);
-  if (!name) {
-    return null;
-  }
-
-  const subtitle = getNonEmptyString(row.subtitle);
-  return subtitle ? `${name} - ${subtitle}` : name;
-}
-
-function hasPendingConnectionDisambiguation(history: Array<{ role: "user" | "assistant"; content: string }>): boolean {
-  for (let index = history.length - 1; index >= 0; index -= 1) {
-    const message = history[index];
-    if (message.role !== "assistant") {
-      continue;
-    }
-    return (
-      /which one did you mean\?/i.test(message.content) &&
-      /\[ref:\s*(member|alumni):[^\]\s]+\]/i.test(message.content)
-    );
-  }
-  return false;
-}
-
-function looksLikeConnectionDisambiguationReply(message: string): boolean {
-  const normalized = message.trim().toLowerCase();
-  if (!normalized || normalized.length > 120) {
-    return false;
-  }
-  if (/\n/.test(normalized)) {
-    return false;
-  }
-  if (
-    /^(?:the\s+)?(?:first|second|third|fourth|fifth|last)\b/.test(normalized) ||
-    /^(?:option\s+)?\d+\b/.test(normalized)
-  ) {
-    return true;
-  }
-  return normalized.split(/\s+/).length <= 8;
-}
-
-// Walk `successfulToolResults` and add any value found at a literal
-// `phone_number` key into `owned`. Unknown-shaped free-text fields are
-// deliberately ignored so a poisoned tool blob cannot widen the allowlist.
-export function collectPhoneNumberFields(value: unknown, owned: Set<string>, depth = 0): void {
-  if (depth > 8 || value == null) return;
-  if (typeof value === "string") return;
-  if (Array.isArray(value)) {
-    for (const item of value) collectPhoneNumberFields(item, owned, depth + 1);
-    return;
-  }
-  if (typeof value !== "object") return;
-  for (const [key, val] of Object.entries(value as Record<string, unknown>)) {
-    if (key === "phone_number" && typeof val === "string" && val.trim().length > 0) {
-      owned.add(val.trim().toLowerCase());
-      continue;
-    }
-    collectPhoneNumberFields(val, owned, depth + 1);
-  }
-}
-
-export function formatSuggestConnectionsResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as SuggestConnectionDisplayPayload;
-  const state = getNonEmptyString(payload.state);
-
-  if (!state) {
-    return null;
-  }
-
-  if (state === "not_found") {
-    return "I couldn't find that person in the organization's member or alumni data. Please share a narrower identifier like a full name or email.";
-  }
-
-  if (state === "ambiguous") {
-    const options = Array.isArray(payload.disambiguation_options)
-      ? payload.disambiguation_options
-          .map((option) => {
-            if (!option || typeof option !== "object") return null;
-            const row = option as {
-              name?: unknown;
-              subtitle?: unknown;
-              person_type?: unknown;
-              person_id?: unknown;
-            };
-            const display = formatDisplayRow(row);
-            if (!display) return null;
-            const personType = getNonEmptyString(row.person_type);
-            const personId = getNonEmptyString(row.person_id);
-            if (personType && personId && (personType === "member" || personType === "alumni")) {
-              // Machine-parseable tail: lets pass-2 prompt re-call the tool
-              // with stable id on the user's next turn. User-facing prefix
-              // stays identical.
-              return `${display} [ref: ${personType}:${personId}]`;
-            }
-            return display;
-          })
-          .filter((option): option is string => Boolean(option))
-      : [];
-
-    if (options.length === 0) {
-      return null;
-    }
-
-    return `I found multiple matches. Which one did you mean?\n${options
-      .map((option) => `- ${option}`)
-      .join("\n")}`;
-  }
-
-  const sourceName = getNonEmptyString(payload.source_person?.name);
-  if (!sourceName) {
-    return null;
-  }
-
-  if (state === "no_suggestions") {
-    return `I found ${sourceName}, but there isn't enough strong professional overlap yet to recommend specific connections within the organization.`;
-  }
-
-  if (state !== "resolved" || !Array.isArray(payload.suggestions)) {
-    return null;
-  }
-
-  const suggestions = payload.suggestions
-    .map((suggestion) => {
-      if (!suggestion || typeof suggestion !== "object") {
-        return null;
-      }
-
-      const displayLine = formatDisplayRow(suggestion as { name?: unknown; subtitle?: unknown });
-      if (!displayLine) {
-        return null;
-      }
-
-      const reasons = Array.isArray((suggestion as SuggestConnectionDisplayRow).reasons)
-        ? ((suggestion as SuggestConnectionDisplayRow).reasons as SuggestConnectionDisplayReason[])
-            .map((reason) => getNonEmptyString(reason?.label))
-            .filter((label): label is string => Boolean(label))
-        : [];
-
-      if (reasons.length === 0) {
-        return null;
-      }
-
-      return { displayLine, reasons };
-    })
-    .filter(
-      (
-        suggestion
-      ): suggestion is {
-        displayLine: string;
-        reasons: string[];
-      } => Boolean(suggestion)
-    )
-    .slice(0, 3);
-
-  if (suggestions.length === 0) {
-    return null;
-  }
-
-  const lines = [`Top connections for ${sourceName}`];
-  for (const [index, suggestion] of suggestions.entries()) {
-    lines.push(`${index + 1}. ${suggestion.displayLine}`);
-    lines.push(`Why: ${suggestion.reasons.join(", ")}`);
-  }
-
-  return lines.join("\n");
-}
-
-interface SuggestMentorsDisplayPayload {
-  state?: unknown;
-  mentee?: { name?: unknown } | null;
-  suggestions?: unknown;
-  disambiguation_options?: unknown;
-}
-
-interface SuggestMentorsDisplaySuggestion {
-  mentor?: { name?: unknown; subtitle?: unknown } | null;
-  reasons?: Array<{ label?: unknown; value?: unknown }>;
-}
-
-interface DonationAnalyticsDisplayPayload {
-  window_days?: unknown;
-  totals?: {
-    successful_donation_count?: unknown;
-    successful_amount_cents?: unknown;
-    average_successful_amount_cents?: unknown;
-    largest_successful_amount_cents?: unknown;
-    latest_successful_donation_at?: unknown;
-    status_counts?: {
-      succeeded?: unknown;
-      failed?: unknown;
-      pending?: unknown;
-    } | null;
-  } | null;
-  trend?: unknown;
-  top_purposes?: unknown;
-}
-
-interface ListAvailableMentorsDisplayPayload {
-  state?: unknown;
-  total_available?: unknown;
-  mentors?: unknown;
-}
-
-interface ListAvailableMentorsDisplayRow {
-  mentor?: { name?: unknown; subtitle?: unknown } | null;
-  open_slots?: unknown;
-  current_mentee_count?: unknown;
-  max_mentees?: unknown;
-  sports?: unknown;
-  positions?: unknown;
-}
-
-function formatSuggestMentorsResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") return null;
-
-  const payload = data as SuggestMentorsDisplayPayload;
-  const state = getNonEmptyString(payload.state);
-  if (!state) return null;
-
-  if (state === "unauthorized") {
-    return "Mentor suggestions are currently available to admins only.";
-  }
-
-  if (state === "not_found") {
-    return "I couldn't find that person in the organization. Please share a full name or email.";
-  }
-
-  if (state === "ambiguous") {
-    const options = Array.isArray(payload.disambiguation_options)
-      ? payload.disambiguation_options
-          .map((option) =>
-            option && typeof option === "object"
-              ? formatDisplayRow(option as { name?: unknown; subtitle?: unknown })
-              : null
-          )
-          .filter((option): option is string => Boolean(option))
-      : [];
-
-    if (options.length === 0) return null;
-
-    return `I found multiple matches. Which one did you mean?\n${options
-      .map((option) => `- ${option}`)
-      .join("\n")}`;
-  }
-
-  const menteeName = getNonEmptyString(payload.mentee?.name);
-  if (!menteeName) return null;
-
-  if (state === "no_suggestions") {
-    return `I found ${menteeName}, but there are no eligible mentors matching their preferences right now.`;
-  }
-
-  if (state !== "resolved" || !Array.isArray(payload.suggestions)) return null;
-
-  const suggestions = (payload.suggestions as SuggestMentorsDisplaySuggestion[])
-    .map((s) => {
-      if (!s || typeof s !== "object") return null;
-      const name = getNonEmptyString(s.mentor?.name);
-      if (!name) return null;
-
-      const subtitle = getNonEmptyString(s.mentor?.subtitle);
-      const displayLine = subtitle ? `${name} — ${subtitle}` : name;
-
-      const reasons = Array.isArray(s.reasons)
-        ? s.reasons
-            .map((r) => {
-              const label = getNonEmptyString(r?.label);
-              if (!label) return null;
-              const value = r?.value;
-              return value != null && value !== "" ? `${label}: ${value}` : label;
-            })
-            .filter((r): r is string => Boolean(r))
-        : [];
-
-      if (reasons.length === 0) return null;
-      return { displayLine, reasons };
-    })
-    .filter(
-      (s): s is { displayLine: string; reasons: string[] } => Boolean(s)
-    )
-    .slice(0, 5);
-
-  if (suggestions.length === 0) return null;
-
-  const lines = [`Top mentors for ${menteeName}`];
-  for (const [index, suggestion] of suggestions.entries()) {
-    lines.push(`${index + 1}. ${suggestion.displayLine}`);
-    lines.push(`   Why: ${suggestion.reasons.join(", ")}`);
-  }
-
-  return lines.join("\n");
-}
-
-function formatDonationAnalyticsResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as DonationAnalyticsDisplayPayload;
-  const totals = payload.totals;
-  if (!totals || typeof totals !== "object") {
-    return null;
-  }
-
-  const windowDays =
-    typeof payload.window_days === "number" ? payload.window_days : null;
-  const lines = [
-    `Donation analytics${windowDays ? ` (${windowDays}-day window)` : ""}`,
-  ];
-
-  if (typeof totals.successful_donation_count === "number") {
-    lines.push(`- Successful donations: ${totals.successful_donation_count}`);
-  }
-  if (typeof totals.successful_amount_cents === "number") {
-    lines.push(`- Raised: $${(totals.successful_amount_cents / 100).toFixed(0)}`);
-  }
-  if (typeof totals.average_successful_amount_cents === "number") {
-    lines.push(
-      `- Average successful donation: $${(totals.average_successful_amount_cents / 100).toFixed(0)}`
-    );
-  }
-  if (typeof totals.largest_successful_amount_cents === "number") {
-    lines.push(
-      `- Largest successful donation: $${(totals.largest_successful_amount_cents / 100).toFixed(0)}`
-    );
-  }
-
-  if (totals.status_counts && typeof totals.status_counts === "object") {
-    const statusSummary = [
-      typeof totals.status_counts.succeeded === "number"
-        ? `${totals.status_counts.succeeded} succeeded`
-        : null,
-      typeof totals.status_counts.pending === "number"
-        ? `${totals.status_counts.pending} pending`
-        : null,
-      typeof totals.status_counts.failed === "number"
-        ? `${totals.status_counts.failed} failed`
-        : null,
-    ].filter((value): value is string => Boolean(value));
-
-    if (statusSummary.length > 0) {
-      lines.push(`- Status mix: ${statusSummary.join(" - ")}`);
-    }
-  }
-
-  const latestSuccessfulDonationAt = formatIsoDate(totals.latest_successful_donation_at);
-  if (latestSuccessfulDonationAt) {
-    lines.push(`- Latest successful donation: ${latestSuccessfulDonationAt}`);
-  }
-
-  const topPurposes = Array.isArray(payload.top_purposes)
-    ? payload.top_purposes
-        .map((row) => {
-          if (!row || typeof row !== "object") return null;
-          const purpose = getNonEmptyString((row as { purpose?: unknown }).purpose);
-          if (!purpose) return null;
-
-          const parts = [
-            typeof (row as { donation_count?: unknown }).donation_count === "number"
-              ? `${(row as { donation_count: number }).donation_count} donations`
-              : null,
-            typeof (row as { amount_cents?: unknown }).amount_cents === "number"
-              ? `$${(((row as { amount_cents: number }).amount_cents) / 100).toFixed(0)}`
-              : null,
-          ].filter((value): value is string => Boolean(value));
-
-          return `- ${purpose}${parts.length > 0 ? ` - ${parts.join(" - ")}` : ""}`;
-        })
-        .filter((row): row is string => Boolean(row))
-        .slice(0, 5)
-    : [];
-
-  if (topPurposes.length > 0) {
-    lines.push("Top purposes");
-    lines.push(...topPurposes);
-  }
-
-  const trendRows = Array.isArray(payload.trend)
-    ? payload.trend
-        .map((row) => {
-          if (!row || typeof row !== "object") return null;
-          const label = getNonEmptyString((row as { bucket_label?: unknown }).bucket_label);
-          if (!label) return null;
-
-          const parts = [
-            typeof (row as { donation_count?: unknown }).donation_count === "number"
-              ? `${(row as { donation_count: number }).donation_count} donations`
-              : null,
-            typeof (row as { amount_cents?: unknown }).amount_cents === "number"
-              ? `$${(((row as { amount_cents: number }).amount_cents) / 100).toFixed(0)}`
-              : null,
-          ].filter((value): value is string => Boolean(value));
-
-          return `- ${label}${parts.length > 0 ? ` - ${parts.join(" - ")}` : ""}`;
-        })
-        .filter((row): row is string => Boolean(row))
-        .slice(0, 8)
-    : [];
-
-  if (trendRows.length > 0) {
-    lines.push("Trend");
-    lines.push(...trendRows);
-  }
-
-  return lines.length > 1 ? lines.join("\n") : null;
-}
-
-function formatListAvailableMentorsResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") return null;
-
-  const payload = data as ListAvailableMentorsDisplayPayload;
-  const state = getNonEmptyString(payload.state);
-  if (!state) return null;
-
-  if (state === "no_results") {
-    return "There are no mentors currently available for new mentees right now.";
-  }
-
-  if (state !== "resolved" || !Array.isArray(payload.mentors)) return null;
-
-  const totalAvailable =
-    typeof payload.total_available === "number" ? payload.total_available : null;
-
-  const mentors = (payload.mentors as ListAvailableMentorsDisplayRow[])
-    .map((row) => {
-      const name = getNonEmptyString(row.mentor?.name);
-      if (!name) return null;
-
-      const subtitle = getNonEmptyString(row.mentor?.subtitle);
-      const displayLine = subtitle ? `${name} — ${subtitle}` : name;
-      const openSlots =
-        typeof row.open_slots === "number" && typeof row.max_mentees === "number"
-          ? `${row.open_slots} open spot${row.open_slots === 1 ? "" : "s"}`
-          : null;
-      const sports = Array.isArray(row.sports)
-        ? row.sports
-            .map((value) => getNonEmptyString(value))
-            .filter((value): value is string => Boolean(value))
-        : [];
-      const positions = Array.isArray(row.positions)
-        ? row.positions
-            .map((value) => getNonEmptyString(value))
-            .filter((value): value is string => Boolean(value))
-        : [];
-      const details = [
-        openSlots,
-        sports.length > 0 ? `Sports: ${sports.join(", ")}` : null,
-        positions.length > 0 ? `Positions: ${positions.join(", ")}` : null,
-      ].filter((value): value is string => Boolean(value));
-
-      return { displayLine, details };
-    })
-    .filter(
-      (mentor): mentor is { displayLine: string; details: string[] } => Boolean(mentor)
-    )
-    .slice(0, 5);
-
-  if (mentors.length === 0) return null;
-
-  const headline =
-    totalAvailable && totalAvailable > mentors.length
-      ? `There are ${totalAvailable} mentors currently available. Here are the top matches by open capacity:`
-      : `There are ${totalAvailable ?? mentors.length} mentors currently available:`;
-  const lines = [headline];
-  for (const [index, mentor] of mentors.entries()) {
-    lines.push(`${index + 1}. ${mentor.displayLine}`);
-    if (mentor.details.length > 0) {
-      lines.push(`   ${mentor.details.join(" • ")}`);
-    }
-  }
-
-  return lines.join("\n");
-}
-
-function formatAnnouncementsResponse(data: unknown): string | null {
-  if (!Array.isArray(data)) {
-    return null;
-  }
-
-  if (data.length === 0) {
-    return "I couldn't find any recent announcements for this organization.";
-  }
-
-  const rows = data
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const announcement = row as AnnouncementDisplayRow;
-      const title = getNonEmptyString(announcement.title);
-      if (!title) {
-        return null;
-      }
-
-      const metadata: string[] = [];
-      const publishedAt = getNonEmptyString(announcement.published_at);
-      if (publishedAt) {
-        metadata.push(publishedAt.slice(0, 10));
-      }
-
-      const audience = getNonEmptyString(announcement.audience);
-      if (audience) {
-        metadata.push(`audience: ${audience.replace(/_/g, " ")}`);
-      }
-
-      if (announcement.is_pinned === true) {
-        metadata.push("pinned");
-      }
-
-      const preview = getNonEmptyString(announcement.body_preview);
-      return {
-        title,
-        metadata,
-        preview,
-      };
-    })
-    .filter(
-      (
-        row
-      ): row is {
-        title: string;
-        metadata: string[];
-        preview: string | null;
-      } => Boolean(row)
-    )
-    .slice(0, 5);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  const lines = ["Recent announcements"];
-  for (const row of rows) {
-    lines.push(`- ${row.title}${row.metadata.length > 0 ? ` - ${row.metadata.join(" - ")}` : ""}`);
-    if (row.preview) {
-      lines.push(`  Preview: ${row.preview}`);
-    }
-  }
-
-  return lines.join("\n");
-}
-
-function formatEventsResponse(data: unknown): string | null {
-  if (!Array.isArray(data)) {
-    return null;
-  }
-
-  if (data.length === 0) {
-    return "I couldn't find any matching events for this organization.";
-  }
-
-  const rows = data
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const title = getNonEmptyString((row as { title?: unknown }).title);
-      if (!title) {
-        return null;
-      }
-
-      const metadata = [
-        formatIsoDate((row as { start_date?: unknown }).start_date),
-        getNonEmptyString((row as { location?: unknown }).location),
-      ].filter((value): value is string => Boolean(value));
-      const description = getNonEmptyString((row as { description?: unknown }).description);
-
-      return { title, metadata, description };
-    })
-    .filter(
-      (
-        row
-      ): row is {
-        title: string;
-        metadata: string[];
-        description: string | null;
-      } => Boolean(row)
-    )
-    .slice(0, 5);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  const lines = ["Matching events"];
-  for (const row of rows) {
-    lines.push(`- ${row.title}${row.metadata.length > 0 ? ` - ${row.metadata.join(" - ")}` : ""}`);
-    if (row.description) {
-      lines.push(`  Details: ${row.description}`);
-    }
-  }
-
-  return lines.join("\n");
-}
-
-function formatDiscussionsResponse(data: unknown): string | null {
-  if (!Array.isArray(data)) {
-    return null;
-  }
-
-  if (data.length === 0) {
-    return "I couldn't find any recent discussion threads for this organization.";
-  }
-
-  const rows = data
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const title = getNonEmptyString((row as { title?: unknown }).title);
-      if (!title) {
-        return null;
-      }
-
-      const metadata = [
-        formatIsoDate((row as { created_at?: unknown }).created_at),
-        typeof (row as { comment_count?: unknown }).comment_count === "number"
-          ? `${(row as { comment_count: number }).comment_count} comments`
-          : null,
-      ].filter((value): value is string => Boolean(value));
-      const preview = getNonEmptyString((row as { body_preview?: unknown }).body_preview);
-
-      return { title, metadata, preview };
-    })
-    .filter(
-      (
-        row
-      ): row is {
-        title: string;
-        metadata: string[];
-        preview: string | null;
-      } => Boolean(row)
-    )
-    .slice(0, 5);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  const lines = ["Recent discussions"];
-  for (const row of rows) {
-    lines.push(`- ${row.title}${row.metadata.length > 0 ? ` - ${row.metadata.join(" - ")}` : ""}`);
-    if (row.preview) {
-      lines.push(`  Preview: ${row.preview}`);
-    }
-  }
-
-  return lines.join("\n");
-}
-
-function formatJobPostingsResponse(data: unknown): string | null {
-  if (!Array.isArray(data)) {
-    return null;
-  }
-
-  if (data.length === 0) {
-    return "I couldn't find any active job postings for this organization.";
-  }
-
-  const rows = data
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const title = getNonEmptyString((row as { title?: unknown }).title);
-      if (!title) {
-        return null;
-      }
-
-      const metadata = [
-        getNonEmptyString((row as { company?: unknown }).company),
-        getNonEmptyString((row as { location?: unknown }).location),
-        getNonEmptyString((row as { job_type?: unknown }).job_type),
-      ].filter((value): value is string => Boolean(value));
-      const preview = getNonEmptyString(
-        (row as { description_preview?: unknown }).description_preview
-      );
-
-      return { title, metadata, preview };
-    })
-    .filter(
-      (
-        row
-      ): row is {
-        title: string;
-        metadata: string[];
-        preview: string | null;
-      } => Boolean(row)
-    )
-    .slice(0, 5);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  const lines = ["Active job postings"];
-  for (const row of rows) {
-    lines.push(`- ${row.title}${row.metadata.length > 0 ? ` - ${row.metadata.join(" - ")}` : ""}`);
-    if (row.preview) {
-      lines.push(`  Preview: ${row.preview}`);
-    }
-  }
-
-  return lines.join("\n");
-}
-
-function formatOrgStatsResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as {
-    active_members?: unknown;
-    alumni?: unknown;
-    parents?: unknown;
-    upcoming_events?: unknown;
-    donations?: {
-      total_amount_cents?: unknown;
-      donation_count?: unknown;
-      last_donation_at?: unknown;
-    } | null;
-  };
-
-  const lines = ["Organization snapshot"];
-
-  if (typeof payload.active_members === "number") {
-    lines.push(`- Active members: ${payload.active_members}`);
-  }
-  if (typeof payload.alumni === "number") {
-    lines.push(`- Alumni: ${payload.alumni}`);
-  }
-  if (typeof payload.parents === "number") {
-    lines.push(`- Parents: ${payload.parents}`);
-  }
-  if (typeof payload.upcoming_events === "number") {
-    lines.push(`- Upcoming events: ${payload.upcoming_events}`);
-  }
-
-  if (payload.donations && typeof payload.donations === "object") {
-    const donationSummary: string[] = [];
-    if (typeof payload.donations.donation_count === "number") {
-      donationSummary.push(`${payload.donations.donation_count} donations`);
-    }
-    if (typeof payload.donations.total_amount_cents === "number") {
-      donationSummary.push(`$${(payload.donations.total_amount_cents / 100).toFixed(0)} raised`);
-    }
-    const lastDonationDate = formatIsoDate(payload.donations.last_donation_at);
-    if (lastDonationDate) {
-      donationSummary.push(`last donation ${lastDonationDate}`);
-    }
-
-    if (donationSummary.length > 0) {
-      lines.push(`- Donations: ${donationSummary.join(" - ")}`);
-    }
-  }
-
-  return lines.length > 1 ? lines.join("\n") : null;
-}
-
-function formatMemberRole(value: unknown): string | null {
-  const role = getNonEmptyString(value);
-  if (!role) {
-    return null;
-  }
-
-  switch (role) {
-    case "admin":
-      return "Admin";
-    case "active_member":
-      return "Active Member";
-    case "alumni":
-      return "Alumni";
-    case "parent":
-      return "Parent";
-    default:
-      return role
-        .split(/[_\s]+/)
-        .filter(Boolean)
-        .map((part) => part.charAt(0).toUpperCase() + part.slice(1))
-        .join(" ");
-  }
-}
-
-function formatMembersResponse(data: unknown): string | null {
-  if (!Array.isArray(data)) {
-    return null;
-  }
-
-  if (data.length === 0) {
-    return "I couldn't find any active members for this organization.";
-  }
-
-  const rows = data
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const email = getNonEmptyString((row as { email?: unknown }).email);
-      const roleLabel = formatMemberRole((row as { role?: unknown }).role);
-      const addedDate = formatIsoDate((row as { created_at?: unknown }).created_at);
-      const name = getNonEmptyString((row as { name?: unknown }).name);
-
-      const label = name
-        ? `${name}${roleLabel ? ` (${roleLabel})` : ""}`
-        : email
-          ? roleLabel === "Admin"
-            ? "Email-only admin account"
-            : "Email-only member account"
-          : null;
-
-      if (!label) {
-        return null;
-      }
-
-      const metadata = [email, addedDate ? `added ${addedDate}` : null].filter(
-        (value): value is string => Boolean(value)
-      );
-
-      return `- ${label}${metadata.length > 0 ? ` - ${metadata.join(" - ")}` : ""}`;
-    })
-    .filter((row): row is string => Boolean(row))
-    .slice(0, 5);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  return ["Recent active members", ...rows].join("\n");
-}
-
-function formatAlumniResponse(data: unknown): string | null {
-  if (!Array.isArray(data)) {
-    return null;
-  }
-
-  if (data.length === 0) {
-    return "I couldn't find any alumni for this organization.";
-  }
-
-  const rows = data
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const name = getNonEmptyString((row as { name?: unknown }).name);
-      if (!name) {
-        return null;
-      }
-
-      const gradYear = typeof (row as { graduation_year?: unknown }).graduation_year === "number"
-        ? `class of ${(row as { graduation_year: number }).graduation_year}`
-        : null;
-      const company = getNonEmptyString((row as { current_company?: unknown }).current_company);
-      const city = getNonEmptyString((row as { current_city?: unknown }).current_city);
-      const title = getNonEmptyString((row as { title?: unknown }).title);
-
-      const metadata = [gradYear, company, city].filter(
-        (value): value is string => Boolean(value)
-      );
-      const suffix = title ? ` (${title})` : "";
-
-      return `- ${name}${metadata.length > 0 ? ` - ${metadata.join(" - ")}` : ""}${suffix}`;
-    })
-    .filter((row): row is string => Boolean(row))
-    .slice(0, 10);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  return ["Alumni", ...rows].join("\n");
-}
-
-function formatEnterpriseAlumniResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as {
-    results?: unknown;
-    total?: unknown;
-  };
-
-  if (!Array.isArray(payload.results)) {
-    return null;
-  }
-
-  if (payload.results.length === 0) {
-    return "I couldn't find any matching alumni across the enterprise.";
-  }
-
-  const rows = payload.results
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const name = getNonEmptyString((row as { name?: unknown }).name);
-      if (!name) {
-        return null;
-      }
-
-      const organizationName = getNonEmptyString(
-        (row as { organization_name?: unknown }).organization_name,
-      );
-      const gradYear =
-        typeof (row as { graduation_year?: unknown }).graduation_year === "number"
-          ? `class of ${(row as { graduation_year: number }).graduation_year}`
-          : null;
-      const company = getNonEmptyString((row as { current_company?: unknown }).current_company);
-      const city = getNonEmptyString((row as { current_city?: unknown }).current_city);
-      const title = getNonEmptyString((row as { title?: unknown }).title);
-
-      const metadata = [
-        organizationName,
-        gradYear,
-        company,
-        city,
-      ].filter((value): value is string => Boolean(value));
-
-      return `- ${name}${metadata.length > 0 ? ` - ${metadata.join(" - ")}` : ""}${title ? ` (${title})` : ""}`;
-    })
-    .filter((row): row is string => Boolean(row))
-    .slice(0, 10);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  const total =
-    typeof payload.total === "number" ? ` (${payload.total} total)` : "";
-
-  return [`Enterprise alumni${total}`, ...rows].join("\n");
-}
-
-function formatEnterpriseStatsResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as {
-    total_count?: unknown;
-    org_stats?: unknown;
-    top_industries?: unknown;
-  };
-
-  const lines = ["Enterprise alumni snapshot"];
-
-  if (typeof payload.total_count === "number") {
-    lines.push(`- Total alumni: ${payload.total_count}`);
-  }
-
-  if (Array.isArray(payload.org_stats) && payload.org_stats.length > 0) {
-    const orgSummary = payload.org_stats
-      .map((row) => {
-        if (!row || typeof row !== "object") {
-          return null;
-        }
-        const name = getNonEmptyString((row as { name?: unknown }).name);
-        const count =
-          typeof (row as { count?: unknown }).count === "number"
-            ? (row as { count: number }).count
-            : null;
-        if (!name || count == null) {
-          return null;
-        }
-        return `${name} (${count})`;
-      })
-      .filter((value): value is string => Boolean(value))
-      .slice(0, 5);
-
-    if (orgSummary.length > 0) {
-      lines.push(`- Org counts: ${orgSummary.join(", ")}`);
-    }
-  }
-
-  if (Array.isArray(payload.top_industries) && payload.top_industries.length > 0) {
-    const industries = payload.top_industries
-      .map((row) => {
-        if (!row || typeof row !== "object") {
-          return null;
-        }
-        const name = getNonEmptyString((row as { name?: unknown }).name);
-        const count =
-          typeof (row as { count?: unknown }).count === "number"
-            ? (row as { count: number }).count
-            : null;
-        if (!name || count == null) {
-          return null;
-        }
-        return `${name} (${count})`;
-      })
-      .filter((value): value is string => Boolean(value))
-      .slice(0, 5);
-
-    if (industries.length > 0) {
-      lines.push(`- Top industries: ${industries.join(", ")}`);
-    }
-  }
-
-  return lines.length > 1 ? lines.join("\n") : null;
-}
-
-function formatEnterpriseQuotaResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as {
-    alumni?: { used?: unknown; limit?: unknown; remaining?: unknown } | null;
-    sub_orgs?: {
-      total?: unknown;
-      enterprise_managed_total?: unknown;
-      free_limit?: unknown;
-      free_remaining?: unknown;
-      configured_limit?: unknown;
-      configured_remaining?: unknown;
-    } | null;
-  };
-
-  const lines = ["Enterprise quota"];
-
-  if (payload.alumni && typeof payload.alumni === "object") {
-    const used =
-      typeof payload.alumni.used === "number" ? payload.alumni.used : null;
-    const limit =
-      typeof payload.alumni.limit === "number" ? payload.alumni.limit : null;
-    const remaining =
-      typeof payload.alumni.remaining === "number" ? payload.alumni.remaining : null;
-
-    if (used != null && limit != null) {
-      lines.push(`- Alumni seats: ${used}/${limit} used`);
-    }
-    if (remaining != null) {
-      lines.push(`- Alumni seats remaining: ${remaining}`);
-    }
-  }
-
-  if (payload.sub_orgs && typeof payload.sub_orgs === "object") {
-    const total =
-      typeof payload.sub_orgs.total === "number" ? payload.sub_orgs.total : null;
-    const enterpriseManagedTotal =
-      typeof payload.sub_orgs.enterprise_managed_total === "number"
-        ? payload.sub_orgs.enterprise_managed_total
-        : null;
-    const freeLimit =
-      typeof payload.sub_orgs.free_limit === "number" ? payload.sub_orgs.free_limit : null;
-    const freeRemaining =
-      typeof payload.sub_orgs.free_remaining === "number"
-        ? payload.sub_orgs.free_remaining
-        : null;
-    const configuredLimit =
-      typeof payload.sub_orgs.configured_limit === "number"
-        ? payload.sub_orgs.configured_limit
-        : null;
-    const configuredRemaining =
-      typeof payload.sub_orgs.configured_remaining === "number"
-        ? payload.sub_orgs.configured_remaining
-        : null;
-
-    if (total != null) {
-      lines.push(`- Managed orgs: ${total}`);
-    }
-    if (enterpriseManagedTotal != null) {
-      lines.push(`- Enterprise-managed org seats in use: ${enterpriseManagedTotal}`);
-    }
-    if (freeLimit != null) {
-      lines.push(`- Free sub-org slots included: ${freeLimit}`);
-    }
-    if (freeRemaining != null) {
-      lines.push(`- Free sub-org slots remaining: ${freeRemaining}`);
-    }
-    if (configuredLimit != null) {
-      lines.push(`- Configured sub-org seat limit: ${configuredLimit}`);
-    }
-    if (configuredRemaining != null) {
-      lines.push(`- Configured sub-org seats remaining: ${configuredRemaining}`);
-    }
-  }
-
-  return lines.length > 1 ? lines.join("\n") : null;
-}
-
-function formatEnterpriseOrgCapacityResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as {
-    sub_orgs?: {
-      total?: unknown;
-      enterprise_managed_total?: unknown;
-      free_limit?: unknown;
-      free_remaining?: unknown;
-    } | null;
-  };
-
-  if (!payload.sub_orgs || typeof payload.sub_orgs !== "object") {
-    return null;
-  }
-
-  const total =
-    typeof payload.sub_orgs.total === "number" ? payload.sub_orgs.total : null;
-  const enterpriseManagedTotal =
-    typeof payload.sub_orgs.enterprise_managed_total === "number"
-      ? payload.sub_orgs.enterprise_managed_total
-      : null;
-  const freeLimit =
-    typeof payload.sub_orgs.free_limit === "number" ? payload.sub_orgs.free_limit : null;
-  const freeRemaining =
-    typeof payload.sub_orgs.free_remaining === "number"
-      ? payload.sub_orgs.free_remaining
-      : null;
-
-  const lines = ["Enterprise managed-org capacity"];
-  if (total != null) {
-    lines.push(`- Managed orgs: ${total}`);
-  }
-  if (enterpriseManagedTotal != null) {
-    lines.push(`- Enterprise-managed org seats in use: ${enterpriseManagedTotal}`);
-  }
-  if (freeLimit != null) {
-    lines.push(`- Free sub-org slots included: ${freeLimit}`);
-  }
-  if (freeRemaining != null) {
-    lines.push(`- Free sub-org slots remaining: ${freeRemaining}`);
-  }
-
-  return lines.length > 1 ? lines.join("\n") : null;
-}
-
-function formatManagedOrgsResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as { organizations?: unknown; total?: unknown };
-  if (!Array.isArray(payload.organizations)) {
-    return null;
-  }
-
-  if (payload.organizations.length === 0) {
-    return "I couldn't find any organizations managed by this enterprise.";
-  }
-
-  const rows = payload.organizations
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const name = getNonEmptyString((row as { name?: unknown }).name);
-      if (!name) {
-        return null;
-      }
-
-      const slug = getNonEmptyString((row as { slug?: unknown }).slug);
-      const relationshipType = getNonEmptyString(
-        (row as { enterprise_relationship_type?: unknown }).enterprise_relationship_type,
-      );
-      const adoptedAt = formatIsoDate(
-        (row as { enterprise_adopted_at?: unknown }).enterprise_adopted_at,
-      );
-
-      const metadata = [slug, relationshipType, adoptedAt].filter(
-        (value): value is string => Boolean(value),
-      );
-
-      return `- ${name}${metadata.length > 0 ? ` - ${metadata.join(" - ")}` : ""}`;
-    })
-    .filter((row): row is string => Boolean(row))
-    .slice(0, 10);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  const total =
-    typeof payload.total === "number" ? ` (${payload.total} total)` : "";
-
-  return [`Managed organizations${total}`, ...rows].join("\n");
-}
-
-function formatAuditEventsResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as { events?: unknown; total?: unknown };
-  if (!Array.isArray(payload.events)) {
-    return null;
-  }
-
-  if (payload.events.length === 0) {
-    return "I couldn't find any recent enterprise audit events.";
-  }
-
-  const rows = payload.events
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const source = getNonEmptyString((row as { source?: unknown }).source);
-      const action = getNonEmptyString((row as { action?: unknown }).action);
-      const when = formatIsoDate((row as { created_at?: unknown }).created_at);
-      const actor = getNonEmptyString(
-        (row as { actor_email_redacted?: unknown }).actor_email_redacted,
-      );
-      const targetType = getNonEmptyString((row as { target_type?: unknown }).target_type);
-      const status = getNonEmptyString((row as { status?: unknown }).status);
-
-      const parts = [action ?? source ?? "event"];
-      if (when) parts.push(when);
-      if (actor) parts.push(`by ${actor}`);
-      if (targetType) parts.push(`target: ${targetType}`);
-      if (status) parts.push(`status: ${status}`);
-
-      return `- ${parts.join(" - ")}`;
-    })
-    .filter((row): row is string => Boolean(row))
-    .slice(0, 25);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  const total =
-    typeof payload.total === "number" ? ` (${payload.total} total)` : "";
-
-  return [`Recent enterprise audit events${total}`, ...rows].join("\n");
-}
-
-interface DonationResponseOptions {
-  hideDonorNames?: boolean;
-}
-
-function formatDonationsResponse(
-  data: unknown,
-  options?: DonationResponseOptions,
-): string | null {
-  if (!Array.isArray(data)) {
-    return null;
-  }
-
-  if (data.length === 0) {
-    return "I couldn't find any donations for this organization.";
-  }
-
-  const rows = data
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const donorName = options?.hideDonorNames
-        ? "Anonymous donor"
-        : (getNonEmptyString((row as { donor_name?: unknown }).donor_name) ?? "Unknown");
-      const amountDollars = typeof (row as { amount_dollars?: unknown }).amount_dollars === "number"
-        ? `$${((row as { amount_dollars: number }).amount_dollars).toFixed(2)}`
-        : null;
-      const status = getNonEmptyString((row as { status?: unknown }).status);
-      const date = formatIsoDate((row as { created_at?: unknown }).created_at);
-      const purpose = getNonEmptyString((row as { purpose?: unknown }).purpose);
-
-      const metadata = [amountDollars, status, date].filter(
-        (value): value is string => Boolean(value)
-      );
-      const suffix = purpose ? ` (${purpose})` : "";
-
-      return `- ${donorName}${metadata.length > 0 ? ` - ${metadata.join(" - ")}` : ""}${suffix}`;
-    })
-    .filter((row): row is string => Boolean(row))
-    .slice(0, 10);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  return ["Donations", ...rows].join("\n");
-}
-
-function formatParentsResponse(data: unknown): string | null {
-  if (!Array.isArray(data)) {
-    return null;
-  }
-
-  if (data.length === 0) {
-    return "I couldn't find any parents for this organization.";
-  }
-
-  const rows = data
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const name = getNonEmptyString((row as { name?: unknown }).name);
-      if (!name) {
-        return null;
-      }
-
-      const relationship = getNonEmptyString((row as { relationship?: unknown }).relationship);
-      const studentName = getNonEmptyString((row as { student_name?: unknown }).student_name);
-
-      const metadata = [
-        relationship,
-        studentName ? `student: ${studentName}` : null,
-      ].filter((value): value is string => Boolean(value));
-
-      return `- ${name}${metadata.length > 0 ? ` - ${metadata.join(" - ")}` : ""}`;
-    })
-    .filter((row): row is string => Boolean(row))
-    .slice(0, 10);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  return ["Parent directory", ...rows].join("\n");
-}
-
-function formatPhilanthropyEventsResponse(data: unknown): string | null {
-  if (!Array.isArray(data)) {
-    return null;
-  }
-
-  if (data.length === 0) {
-    return "I couldn't find any philanthropy events for this organization.";
-  }
-
-  const rows = data
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const title = getNonEmptyString((row as { title?: unknown }).title);
-      if (!title) {
-        return null;
-      }
-
-      const metadata = [
-        formatIsoDate((row as { start_date?: unknown }).start_date),
-        getNonEmptyString((row as { location?: unknown }).location),
-      ].filter((value): value is string => Boolean(value));
-      const description = getNonEmptyString((row as { description?: unknown }).description);
-
-      return { title, metadata, description };
-    })
-    .filter(
-      (
-        row
-      ): row is {
-        title: string;
-        metadata: string[];
-        description: string | null;
-      } => Boolean(row)
-    )
-    .slice(0, 5);
-
-  if (rows.length === 0) {
-    return null;
-  }
-
-  const lines = ["Philanthropy events"];
-  for (const row of rows) {
-    lines.push(`- ${row.title}${row.metadata.length > 0 ? ` - ${row.metadata.join(" - ")}` : ""}`);
-    if (row.description) {
-      lines.push(`  Details: ${row.description}`);
-    }
-  }
-
-  return lines.join("\n");
-}
-
-function formatNavigationTargetsResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as NavigationDisplayPayload;
-  const state = getNonEmptyString(payload.state);
-  const query = getNonEmptyString(payload.query) ?? "that";
-
-  if (!state) {
-    return null;
-  }
-
-  if (state === "not_found") {
-    return `I couldn't find a matching page for "${query}". Try naming the feature, like announcements, members, events, donations, or navigation settings.`;
-  }
-
-  if (state !== "resolved" || !Array.isArray(payload.matches)) {
-    return null;
-  }
-
-  const matches = payload.matches
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const target = row as NavigationDisplayTarget;
-      const label = getNonEmptyString(target.label);
-      const href = getNonEmptyString(target.href);
-      if (!label || !href) {
-        return null;
-      }
-
-      const description = getNonEmptyString(target.description);
-      const kind = getNonEmptyString(target.kind);
-      const manualSteps = Array.isArray(target.manualSteps)
-        ? target.manualSteps.map(getNonEmptyString).filter((value): value is string => Boolean(value))
-        : [];
-      const assistantCanHelpWith = Array.isArray(target.assistantCanHelpWith)
-        ? target.assistantCanHelpWith.map(getNonEmptyString).filter((value): value is string => Boolean(value))
-        : [];
-      const userCanAccess =
-        typeof target.userCanAccess === "boolean" ? target.userCanAccess : true;
-      return { label, href, description, kind, manualSteps, assistantCanHelpWith, userCanAccess };
-    })
-    .filter(
-      (
-        row
-      ): row is {
-        label: string;
-        href: string;
-        description: string | null;
-        kind: string | null;
-        manualSteps: string[];
-        assistantCanHelpWith: string[];
-        userCanAccess: boolean;
-      } => Boolean(row)
-    )
-    .filter((row) => row.userCanAccess)
-    .slice(0, 3);
-
-  if (matches.length === 0) {
-    return null;
-  }
-
-  const escapeLabel = (label: string) => label.replace(/_/g, "\\_");
-
-  const blocks: string[] = [];
-  if (matches.length > 1) {
-    blocks.push(`Best matches for "${query}":`);
-  }
-  for (const match of matches) {
-    const header =
-      matches.length === 1
-        ? `[${escapeLabel(match.label)}](${match.href})`
-        : `- [${escapeLabel(match.label)}](${match.href})`;
-    const lines: string[] = [header];
-    const manualStep = match.manualSteps[0];
-    if (manualStep) {
-      lines.push(`Next: ${manualStep}`);
-    }
-    const assistantHelp = match.assistantCanHelpWith[0];
-    if (assistantHelp) {
-      lines.push(`I can help: ${assistantHelp}`);
-    }
-    blocks.push(lines.join("  \n"));
-  }
-
-  return blocks.join("\n\n");
 }
 
 function getPass1Tools(
@@ -2314,634 +792,6 @@ function getBatchPendingActionsFromToolData(data: unknown) {
 
   return actions.length > 0 ? actions : null;
 }
-
-function formatPrepareJobPostingResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as PendingActionToolPayload;
-  if (payload.state === "invalid_source_url") {
-    return typeof payload.message === "string" && payload.message.length > 0
-      ? `I couldn't read that job posting URL safely. ${payload.message}`
-      : "I couldn't read that job posting URL safely. Please provide the job details directly.";
-  }
-
-  if (payload.state === "missing_fields") {
-    const missingFields = Array.isArray(payload.missing_fields)
-      ? payload.missing_fields.filter((field): field is string => typeof field === "string" && field.length > 0)
-      : [];
-    const sourceWarning =
-      typeof payload.source_warning === "string" && payload.source_warning.length > 0
-        ? payload.source_warning
-        : null;
-
-    if (missingFields.length === 0) {
-      return sourceWarning
-        ? `I couldn't read that job posting URL safely, but I can still draft this job if you share a few more details.`
-        : "I still need a few more job details before I can prepare this posting.";
-    }
-
-    return sourceWarning
-      ? `I couldn't read that job posting URL safely, so I still need: ${missingFields.join(", ")}.`
-      : `I can draft this job, but I still need: ${missingFields.join(", ")}.`;
-  }
-
-  if (payload.state === "needs_confirmation") {
-    return "I drafted the job posting. Review the details below and confirm when you're ready to create it.";
-  }
-
-  return null;
-}
-
-function formatPrepareAnnouncementResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as PendingActionToolPayload;
-  if (payload.state === "missing_fields") {
-    const missingFields = Array.isArray(payload.missing_fields)
-      ? payload.missing_fields.filter((field): field is string => typeof field === "string" && field.length > 0)
-      : [];
-
-    if (missingFields.length === 0) {
-      return "I still need an announcement title before I can prepare this post.";
-    }
-
-    return `I can draft this announcement, but I still need: ${missingFields.join(", ")}.`;
-  }
-
-  if (payload.state === "needs_confirmation") {
-    return "I drafted the announcement. Review the details below and confirm when you're ready to publish it.";
-  }
-
-  return null;
-}
-
-function formatPrepareEnterpriseInviteResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-  const payload = data as PendingActionToolPayload;
-  if (payload.state === "missing_fields") {
-    const missingFields = Array.isArray(payload.missing_fields)
-      ? payload.missing_fields.filter(
-          (field): field is string => typeof field === "string" && field.length > 0,
-        )
-      : [];
-    if (missingFields.length === 0) {
-      return "I can draft this enterprise invite, but I still need more details before I can prepare it.";
-    }
-    return `I can draft this enterprise invite, but I still need: ${missingFields.join(", ")}.`;
-  }
-  if (payload.state === "needs_confirmation") {
-    return "I drafted the enterprise invite. Review the details below and confirm when you're ready to create it.";
-  }
-  return null;
-}
-
-function formatRevokeEnterpriseInviteResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-  const payload = data as PendingActionToolPayload;
-  if (payload.state === "needs_confirmation") {
-    return "I found that enterprise invite. Confirm below to revoke it.";
-  }
-  return null;
-}
-
-function formatPrepareDiscussionThreadResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as PendingActionToolPayload;
-  if (payload.state === "missing_fields") {
-    const missingFields = Array.isArray(payload.missing_fields)
-      ? payload.missing_fields.filter((field): field is string => typeof field === "string" && field.length > 0)
-      : [];
-
-    if (missingFields.length === 0) {
-      return "I still need a discussion title and body before I can prepare this thread.";
-    }
-
-    return `I can draft this discussion, but I still need: ${missingFields.join(", ")}.`;
-  }
-
-  if (payload.state === "needs_confirmation") {
-    return "I drafted the discussion thread. Review the details below and confirm when you're ready to post it.";
-  }
-
-  return null;
-}
-
-function formatPrepareDiscussionReplyResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as PendingActionToolPayload;
-  if (payload.state === "missing_fields") {
-    const clarificationKind = getNonEmptyString(payload.clarification_kind);
-    const requestedThreadTitle = getNonEmptyString(payload.requested_thread_title);
-    const candidateThreadTitles = Array.isArray(payload.candidate_thread_titles)
-      ? payload.candidate_thread_titles.filter(
-          (title): title is string => typeof title === "string" && title.trim().length > 0
-        )
-      : [];
-    const missingFields = Array.isArray(payload.missing_fields)
-      ? payload.missing_fields.filter((field): field is string => typeof field === "string" && field.length > 0)
-      : [];
-
-    if (clarificationKind === "thread_title_required") {
-      return "I can draft this discussion reply, but I still need the thread title before I can prepare it.";
-    }
-
-    if (clarificationKind === "thread_title_ambiguous") {
-      const options =
-        candidateThreadTitles.length > 0
-          ? candidateThreadTitles.join("; ")
-          : "the matching discussion threads";
-      return `I found a few discussion threads that match${
-        requestedThreadTitle ? ` "${requestedThreadTitle}"` : ""
-      }. Tell me which one you mean: ${options}.`;
-    }
-
-    if (clarificationKind === "thread_title_not_found") {
-      if (requestedThreadTitle) {
-        return `I couldn't find a discussion thread titled "${requestedThreadTitle}". Share a more specific thread title and I'll use that.`;
-      }
-      return "I couldn't find that discussion thread. Share a more specific thread title and I'll use that.";
-    }
-
-    if (clarificationKind === "thread_lookup_failed") {
-      return "I couldn't look up that thread right now. Please try again.";
-    }
-
-    if (missingFields.length === 0) {
-      return "I still need the reply body and the target discussion thread title before I can prepare this reply.";
-    }
-
-    const displayFields = missingFields.map((field) =>
-      field === "discussion_thread_id" ? "thread title" : field
-    );
-
-    return `I can draft this discussion reply, but I still need: ${displayFields.join(", ")}.`;
-  }
-
-  if (payload.state === "needs_confirmation") {
-    return "I drafted the discussion reply. Review the details below and confirm when you're ready to post it.";
-  }
-
-  return null;
-}
-
-function formatPrepareChatMessageResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as PendingActionToolPayload;
-  if (payload.state === "missing_fields") {
-    const clarificationKind = getNonEmptyString(payload.clarification_kind);
-    const requestedRecipient = getNonEmptyString(payload.requested_recipient);
-    const candidateRecipients = Array.isArray(payload.candidate_recipients)
-      ? payload.candidate_recipients.filter(
-          (value): value is string => typeof value === "string" && value.trim().length > 0
-        )
-      : [];
-    const missingFields = Array.isArray(payload.missing_fields)
-      ? payload.missing_fields.filter((field): field is string => typeof field === "string" && field.length > 0)
-      : [];
-
-    if (clarificationKind === "recipient_required") {
-      if (missingFields.includes("body")) {
-        return "I can draft that chat message, but I still need who it should go to and the message body.";
-      }
-      return "I can draft that chat message, but I still need to know who should receive it.";
-    }
-
-    if (clarificationKind === "recipient_ambiguous") {
-      const options =
-        candidateRecipients.length > 0 ? candidateRecipients.join("; ") : "the matching members";
-      return `I found a few members that match${
-        requestedRecipient ? ` "${requestedRecipient}"` : ""
-      }. Tell me which one you mean: ${options}.`;
-    }
-
-    if (clarificationKind === "recipient_unavailable") {
-      if (requestedRecipient) {
-        return `I can't send an in-app chat message to "${requestedRecipient}" right now. Pick a different member or choose someone with an active linked account.`;
-      }
-      return "I can't send an in-app chat message to that person right now. Pick a different member or choose someone with an active linked account.";
-    }
-
-    if (missingFields.length === 0) {
-      return "I still need the chat message details before I can prepare it.";
-    }
-
-    const displayFields = missingFields.map((field) =>
-      field === "person_query" ? "recipient" : field
-    );
-    return `I can draft that chat message, but I still need: ${displayFields.join(", ")}.`;
-  }
-
-  if (payload.state === "needs_confirmation") {
-    return "I drafted the chat message. Review the details below and confirm when you're ready to send it.";
-  }
-
-  return null;
-}
-
-function formatChatGroupsResponse(data: unknown): string | null {
-  if (!Array.isArray(data)) {
-    return null;
-  }
-
-  const groups = data
-    .map((row) => {
-      if (!row || typeof row !== "object") {
-        return null;
-      }
-
-      const name = getNonEmptyString((row as { name?: unknown }).name);
-      if (!name) {
-        return null;
-      }
-
-      const role = getNonEmptyString((row as { role?: unknown }).role);
-      return role ? `${name} (${role})` : name;
-    })
-    .filter((row): row is string => Boolean(row))
-    .slice(0, 8);
-
-  if (groups.length === 0) {
-    return "You do not have any active chat groups available right now.";
-  }
-
-  return `You can message these chat groups:\n- ${groups.join("\n- ")}`;
-}
-
-function formatPrepareGroupMessageResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as PendingActionToolPayload;
-  if (payload.state === "missing_fields") {
-    const clarificationKind = getNonEmptyString(payload.clarification_kind);
-    const requestedGroup = getNonEmptyString(payload.requested_group);
-    const candidateGroups = Array.isArray(payload.candidate_groups)
-      ? payload.candidate_groups
-          .map((value) => {
-            if (!value || typeof value !== "object") {
-              return null;
-            }
-
-            return getNonEmptyString((value as { name?: unknown }).name);
-          })
-          .filter((value): value is string => Boolean(value))
-      : [];
-    const missingFields = Array.isArray(payload.missing_fields)
-      ? payload.missing_fields.filter((field): field is string => typeof field === "string" && field.length > 0)
-      : [];
-
-    if (clarificationKind === "group_required") {
-      if (missingFields.includes("body")) {
-        return "I can draft that group message, but I still need which chat group it should go to and the message body.";
-      }
-      return "I can draft that group message, but I still need to know which chat group should receive it.";
-    }
-
-    if (clarificationKind === "group_ambiguous") {
-      const options =
-        candidateGroups.length > 0 ? candidateGroups.join("; ") : "the matching chat groups";
-      return `I found a few chat groups that match${
-        requestedGroup ? ` "${requestedGroup}"` : ""
-      }. Tell me which one you mean: ${options}.`;
-    }
-
-    if (clarificationKind === "group_unavailable") {
-      if (requestedGroup) {
-        return `I can't send an in-app group chat message to "${requestedGroup}" right now. Pick a different chat group or choose one you still belong to.`;
-      }
-      return "I can't send an in-app group chat message there right now. Pick a different chat group or choose one you still belong to.";
-    }
-
-    if (missingFields.length === 0) {
-      return "I still need the group chat message details before I can prepare it.";
-    }
-
-    const displayFields = missingFields.map((field) =>
-      field === "group_name_query" ? "chat group" : field
-    );
-    return `I can draft that group message, but I still need: ${displayFields.join(", ")}.`;
-  }
-
-  if (payload.state === "needs_confirmation") {
-    return "I drafted the group message. Review the details below and confirm when you're ready to send it.";
-  }
-
-  return null;
-}
-
-function formatPrepareEventResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as PendingActionToolPayload;
-  if (payload.state === "missing_fields") {
-    const missingFields = Array.isArray(payload.missing_fields)
-      ? payload.missing_fields.filter((field): field is string => typeof field === "string" && field.length > 0)
-      : [];
-
-    if (missingFields.length === 0) {
-      return "I still need an event title, start date, and start time before I can prepare this event.";
-    }
-
-    return `I can draft this event, but I still need: ${missingFields.join(", ")}.`;
-  }
-
-  if (payload.state === "needs_confirmation") {
-    return "I drafted the event. Review the details below and confirm when you're ready to add it to the calendar.";
-  }
-
-  return null;
-}
-
-function formatPrepareEventsBatchResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as {
-    state?: string;
-    pending_actions?: unknown[];
-    validation_errors?: Array<{ index: number; missing_fields: string[] }>;
-  };
-
-  if (payload.state === "missing_fields") {
-    const errors = payload.validation_errors ?? [];
-    const allMissing = errors.flatMap((e) => e.missing_fields);
-    const unique = [...new Set(allMissing)];
-    if (unique.length === 0) {
-      return "I need more details for these events before I can prepare them.";
-    }
-    return `None of the events are ready yet. I still need: ${unique.join(", ")} for each event.`;
-  }
-
-  if (payload.state === "needs_batch_confirmation") {
-    const count = Array.isArray(payload.pending_actions) ? payload.pending_actions.length : 0;
-    const errorCount = Array.isArray(payload.validation_errors) ? payload.validation_errors.length : 0;
-    let msg = `I drafted ${count} event${count !== 1 ? "s" : ""}. Review the details below and confirm when you're ready.`;
-    if (errorCount > 0) {
-      msg += ` ${errorCount} event${errorCount !== 1 ? "s" : ""} couldn't be prepared — I'll need more details for ${errorCount === 1 ? "that one" : "those"}.`;
-    }
-    return msg;
-  }
-
-  return null;
-}
-
-function formatRevisedPendingEventResponse(data: unknown, count: number): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as {
-    state?: string;
-    validation_errors?: Array<{ index: number; missing_fields: string[] }>;
-  };
-
-  if (payload.state === "missing_fields") {
-    const errors = Array.isArray(payload.validation_errors) ? payload.validation_errors : [];
-    const missingFields = [...new Set(errors.flatMap((error) => error.missing_fields))];
-
-    if (missingFields.length === 0) {
-      return "I updated the drafted schedule, but I still need a few more details before it is ready to confirm again.";
-    }
-
-    return `I updated the drafted schedule, but I still need: ${missingFields.join(", ")} before you can confirm the revised events.`;
-  }
-
-  if (payload.state === "needs_batch_confirmation" || payload.state === "needs_confirmation") {
-    return count === 1
-      ? "I revised the drafted event. Review the updated details below and confirm when you're ready."
-      : "I revised the drafted schedule. Review the updated details below and confirm when you're ready.";
-  }
-
-  return null;
-}
-
-function formatExtractScheduleFileResponse(data: unknown): string | null {
-  if (!data || typeof data !== "object") {
-    return null;
-  }
-
-  const payload = data as {
-    state?: string;
-    pending_actions?: unknown[];
-    validation_errors?: Array<{ index: number; missing_fields: string[] }>;
-    source_file?: unknown;
-  };
-
-  if (payload.state === "no_events_found") {
-    return "I couldn't find any usable events in that schedule file. Try a clearer photo or upload a PDF export if you have one.";
-  }
-
-  if (payload.state === "missing_fields") {
-    const errors = Array.isArray(payload.validation_errors) ? payload.validation_errors : [];
-    const missingFields = [...new Set(errors.flatMap((error) => error.missing_fields))];
-
-    if (missingFields.length === 0) {
-      return "I could read the schedule file, but I need a few more event details before I can prepare anything for confirmation.";
-    }
-
-    return `I could read the schedule file, but I still need: ${missingFields.join(", ")} before I can prepare those events.`;
-  }
-
-  if (payload.state === "needs_batch_confirmation") {
-    const count = Array.isArray(payload.pending_actions) ? payload.pending_actions.length : 0;
-    const skipped = Array.isArray(payload.validation_errors) ? payload.validation_errors.length : 0;
-    let message = `I drafted ${count} event${count === 1 ? "" : "s"} from that schedule file. Review the details below and confirm when you're ready.`;
-    if (skipped > 0) {
-      message += ` ${skipped} event${skipped === 1 ? "" : "s"} still need more details.`;
-    }
-    return message;
-  }
-
-  return null;
-}
-
-function formatDeterministicToolErrorResponse(
-  name: string,
-  error: string,
-  errorCode?: string | null
-): string | null {
-  if (name !== "extract_schedule_pdf") {
-    if (
-      name === "get_enterprise_quota" &&
-      (errorCode === "enterprise_billing_role_required" ||
-        /enterprise owner or billing admin role/i.test(error))
-    ) {
-      return "I can’t access enterprise quota or billing details for your role. Enterprise owners and billing admins can view alumni seat limits and billing quotas.";
-    }
-
-    return null;
-  }
-
-  switch (errorCode) {
-    case "attachment_required":
-      return "I need an uploaded schedule file before I can import anything. Please attach a PDF or schedule image and try again.";
-    case "invalid_attachment_path":
-      return "That uploaded schedule file is no longer valid for this session. Please upload it again.";
-    case "org_context_failed":
-      return "I couldn't load the organization context for that schedule import right now. Please try again.";
-    case "attachment_unavailable":
-      return "I couldn't load that uploaded schedule file. Please re-upload it and try again.";
-    case "image_too_large":
-      return "That schedule image is too large to process. Please upload an image under 2MB or use a PDF instead.";
-    case "image_timeout":
-      return "I wasn't able to extract the schedule from the attached image file because the extraction tool timed out. This can happen with larger or more complex image files. Please re-upload it and I'll try again, or upload a PDF version if you have one.";
-    case "image_unreadable":
-      return "I couldn't read that schedule image. Try a clearer photo, better lighting, or upload a PDF version of the schedule.";
-    case "image_model_misconfigured":
-      return "Schedule image extraction is misconfigured in this environment. Set ZAI_IMAGE_MODEL to a Z.AI vision model like glm-5v-turbo and restart the server.";
-    case "pdf_timeout":
-      return "The attached PDF schedule timed out during extraction. Please try again, or send the event details in text if the PDF keeps failing.";
-    case "pdf_unreadable":
-      return "I couldn't read that PDF schedule. Try re-exporting the PDF or upload a clear image instead.";
-    default:
-      break;
-  }
-
-  if (error === "Unable to read attached schedule image") {
-    return "I couldn't read that schedule image. Try a clearer photo, better lighting, or upload a PDF version of the schedule.";
-  }
-
-  if (error === "Schedule image extraction timed out") {
-    return "I wasn't able to extract the schedule from the attached image file because the extraction tool timed out. This can happen with larger or more complex image files. Please re-upload it and I'll try again, or upload a PDF version if you have one.";
-  }
-
-  if (error === "Schedule PDF extraction timed out") {
-    return "The attached PDF schedule timed out during extraction. Please try again, or send the event details in text if the PDF keeps failing.";
-  }
-
-  if (
-    error ===
-    "Schedule image extraction is misconfigured. Set ZAI_IMAGE_MODEL to a Z.AI vision model such as glm-5v-turbo."
-  ) {
-    return "Schedule image extraction is misconfigured in this environment. Set ZAI_IMAGE_MODEL to a Z.AI vision model like glm-5v-turbo and restart the server.";
-  }
-
-  if (error === "Unable to read attached PDF") {
-    return "I couldn't read that PDF schedule. Try re-exporting the PDF or upload a clear image instead.";
-  }
-
-  return null;
-}
-
-async function resolveHideDonorNamesPreference(
-  serviceSupabase: { from: (table: string) => any },
-  orgId: string,
-): Promise<boolean> {
-  try {
-    const { data, error } = await serviceSupabase
-      .from("organizations")
-      .select("hide_donor_names")
-      .eq("id", orgId)
-      .maybeSingle();
-
-    if (error) {
-      return true;
-    }
-
-    return Boolean((data as { hide_donor_names?: unknown } | null)?.hide_donor_names);
-  } catch {
-    return true;
-  }
-}
-
-function formatDeterministicToolResponse(
-  name: string,
-  data: unknown,
-  options?: DonationResponseOptions,
-): string | null {
-  switch (name) {
-    case "suggest_connections":
-      return formatSuggestConnectionsResponse(data);
-    case "suggest_mentors":
-      return formatSuggestMentorsResponse(data);
-    case "list_available_mentors":
-      return formatListAvailableMentorsResponse(data);
-    case "list_events":
-      return formatEventsResponse(data);
-    case "list_announcements":
-      return formatAnnouncementsResponse(data);
-    case "list_chat_groups":
-      return formatChatGroupsResponse(data);
-    case "list_discussions":
-      return formatDiscussionsResponse(data);
-    case "list_job_postings":
-      return formatJobPostingsResponse(data);
-    case "prepare_announcement":
-      return formatPrepareAnnouncementResponse(data);
-    case "prepare_job_posting":
-      return formatPrepareJobPostingResponse(data);
-    case "prepare_chat_message":
-      return formatPrepareChatMessageResponse(data);
-    case "prepare_group_message":
-      return formatPrepareGroupMessageResponse(data);
-    case "prepare_discussion_reply":
-      return formatPrepareDiscussionReplyResponse(data);
-    case "prepare_discussion_thread":
-      return formatPrepareDiscussionThreadResponse(data);
-    case "prepare_event":
-      return formatPrepareEventResponse(data);
-    case "prepare_events_batch":
-      return formatPrepareEventsBatchResponse(data);
-    case "extract_schedule_pdf":
-      return formatExtractScheduleFileResponse(data);
-    case "get_org_stats":
-      return formatOrgStatsResponse(data);
-    case "get_donation_analytics":
-      return formatDonationAnalyticsResponse(data);
-    case "get_enterprise_stats":
-      return formatEnterpriseStatsResponse(data);
-    case "get_enterprise_quota":
-      return formatEnterpriseQuotaResponse(data);
-    case "get_enterprise_org_capacity":
-      return formatEnterpriseOrgCapacityResponse(data);
-    case "list_members":
-      return formatMembersResponse(data);
-    case "list_alumni":
-      return formatAlumniResponse(data);
-    case "list_enterprise_alumni":
-      return formatEnterpriseAlumniResponse(data);
-    case "list_donations":
-      return formatDonationsResponse(data, options);
-    case "list_managed_orgs":
-      return formatManagedOrgsResponse(data);
-    case "list_enterprise_audit_events":
-      return formatAuditEventsResponse(data);
-    case "prepare_enterprise_invite":
-      return formatPrepareEnterpriseInviteResponse(data);
-    case "revoke_enterprise_invite":
-      return formatRevokeEnterpriseInviteResponse(data);
-    case "list_parents":
-      return formatParentsResponse(data);
-    case "list_philanthropy_events":
-      return formatPhilanthropyEventsResponse(data);
-    case "find_navigation_targets":
-      return formatNavigationTargetsResponse(data);
-    default:
-      return null;
-  }
-}
-
 
 
 const MESSAGE_SAFETY_FALLBACK =
